requirements.txt
streamlit==1.32.0
pandas==2.2.1
sqlalchemy>=1.4.0,<2.0.0
snowflake-sqlalchemy==1.5.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9 